+++ packages/adhere-ui-hooks/es/index.js
@@ -1,2 +1,2 @@
-import useFirst from"./useFirst";import usePrevious from"./usePrevious";import useForceUpdate from"./useForceUpdate";export default{useFirst:useFirst,usePrevious:usePrevious,useForceUpdate:useForceUpdate};
+import useFirst from"./useFirst";import usePrevious from"./usePrevious";import useForceUpdate from"./useForceUpdate";import useSetState from"./useSetState";export default{useFirst:useFirst,usePrevious:usePrevious,useForceUpdate:useForceUpdate,useSetState:useSetState};
 //# sourceMappingURL=index.js.map