+++ packages/adhere-ui-hooks/src/index.ts
@@ -1,5 +1,6 @@
 import useFirst from './useFirst';
 import usePrevious from './usePrevious';
 import useForceUpdate from './useForceUpdate';
+import useSetState from './useSetState';
 
-export default { useFirst, usePrevious, useForceUpdate };
+export default { useFirst, usePrevious, useForceUpdate, useSetState };