+++ packages/adhere-website/src/components/ui/hooks/index.jsx
@@ -1,17 +1,17 @@
 import React, { useEffect } from 'react';
 import { Button } from 'antd';
-import { Hooks } from '@baifendian/adhere';
+import { Hooks, Space } from '@baifendian/adhere';
 
 import PlayGroundPage, { Section, CodeBoxSection } from '@/lib/PlaygroundPage';
 
-const { useFirst, useForceUpdate, usePrevious } = Hooks;
+const { useFirst, useForceUpdate, usePrevious, useSetState } = Hooks;
 
 export default () => {
   function boxPanelConfig() {
     return [
       {
         id: `p1`,
-        name: `基本的使用`,
+        name: `useForceUpdate`,
         mode: 'code',
         scope: { React },
         cardProps: {
@@ -58,6 +58,7 @@ export default () => {
           <Button
             type="primary"
             onClick={() => {
+              // eslint-disable-next-line @typescript-eslint/no-use-before-define
               forceUpdate();
             }}
           >
@@ -65,8 +66,123 @@ export default () => {
           </Button>
         ),
       },
+      {
+        id: `p2`,
+        name: `useSetState`,
+        mode: 'code',
+        scope: { React },
+        cardProps: {
+          description: {
+            title: '基本的使用',
+            info: '基本的使用',
+          },
+        },
+        codeText: `
+  import React, { useEffect } from 'react';
+  import { Button, Space } from 'antd';
+  import { Hooks } from '@baifendian/adhere';
+
+  const { useSetState } = Hooks;
+
+  export default () => {
+    const [count1, setCount1] = useSetState(0);
+
+    const [count2, setCount2] = useSetState(0);
+
+    const [count3, setCount3] = useSetState(0);
+
+    return (
+      <div>
+        <div>
+          <Space.Group direction="horizontal">
+            <Button
+              onClick={() => {
+                setCount1(count1 + 1, () => console.log('count1更新完成'));
+              }}
+            >
+              setCount1
+            </Button>
+            <Button
+              onClick={() => {
+                setCount2(count2 + 1, () => console.log('count2更新完成'));
+              }}
+            >
+              setCount2
+            </Button>
+            <Button
+              onClick={() => {
+                setCount3(
+                  (count) => count + 1,
+                  () => console.log('count3更新完成'),
+                );
+              }}
+            >
+              setCount3
+            </Button>
+          </Space.Group>
+        </div>
+        <div>
+          <p>count1: {count1}</p>
+          <p>count2: {count2}</p>
+          <p>count3: {count3}</p>
+        </div>
+      </div>
+    )
+  }
+        `,
+        type: 'PlayGround',
+        renderChildren: () => (
+          <div>
+            <div>
+              <Space.Group direction="horizontal">
+                <Button
+                  onClick={() => {
+                    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+                    setCount1(count1 + 1, () => console.log('count1更新完成'));
+                  }}
+                >
+                  setCount1
+                </Button>
+                <Button
+                  onClick={() => {
+                    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+                    setCount2(count2 + 1, () => console.log('count2更新完成'));
+                  }}
+                >
+                  setCount2
+                </Button>
+                <Button
+                  onClick={() => {
+                    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+                    setCount3(
+                      (count) => count + 1,
+                      () => console.log('count3更新完成'),
+                    );
+                  }}
+                >
+                  setCount3
+                </Button>
+              </Space.Group>
+            </div>
+            <div>
+              {/* eslint-disable-next-line @typescript-eslint/no-use-before-define */}
+              <p>count1: {count1}</p>
+              {/* eslint-disable-next-line @typescript-eslint/no-use-before-define */}
+              <p>count2: {count2}</p>
+              {/* eslint-disable-next-line @typescript-eslint/no-use-before-define */}
+              <p>count3: {count3}</p>
+            </div>
+          </div>
+        ),
+      },
     ];
   }
+
+  const [count1, setCount1] = useSetState(0);
+
+  const [count2, setCount2] = useSetState(0);
+
+  const [count3, setCount3] = useSetState(0);
 
   const [isFirst, updateFirst] = useFirst();